[Cookie] Move jar to a sub namespace
