+++ src/Ivory/HttpAdapter/Event/Cookie/Jar/AbstractPersistentCookieJar.php
@@ -9,7 +9,10 @@
  * file that was distributed with this source code.
  */
 
-namespace Ivory\HttpAdapter\Event\Cookie;
+namespace Ivory\HttpAdapter\Event\Cookie\Jar;
+
+use Ivory\HttpAdapter\Event\Cookie\CookieFactoryInterface;
+use Ivory\HttpAdapter\Event\Cookie\CookieInterface;
 
 /**
  * Abstract persistent cookie jar.
