+++ src/Ivory/HttpAdapter/Event/Cookie/Jar/CookieJar.php
@@ -9,8 +9,11 @@
  * file that was distributed with this source code.
  */
 
-namespace Ivory\HttpAdapter\Event\Cookie;
+namespace Ivory\HttpAdapter\Event\Cookie\Jar;
 
+use Ivory\HttpAdapter\Event\Cookie\CookieFactoryInterface;
+use Ivory\HttpAdapter\Event\Cookie\CookieFactory;
+use Ivory\HttpAdapter\Event\Cookie\CookieInterface;
 use Ivory\HttpAdapter\Message\InternalRequestInterface;
 use Ivory\HttpAdapter\Message\ResponseInterface;
